simple_user_app/requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
Django==5.1.5
django-redis==5.4.0
djangorestframework==3.15.2
redis==5.2.1
sqlparse==0.5.3
typing_extensions==4.12.2
